update: server side fetch of executive user panel
